Move stories to the component folders. Remove create react app in src folder. Load stories dynamically
