+++ .storybook/config.js
@@ -2,9 +2,12 @@ import { configure, setAddon } from "@storybook/react";
 import staticMarkup from "react-storybook-addon-static-markup";
 
 setAddon(staticMarkup);
+require("./stories.index");
 
+// automatically import all files ending in *.stories.ts
+const req = require.context('../src', true, /.stories.js$/);
 function loadStories() {
-  require("../src/stories.index");
+  req.keys().forEach(filename => req(filename));
 }
 
 configure(loadStories, module);
