+++ src/Accordion/Accordion.scss
@@ -1,5 +1,3 @@
-@import "../_core/_index";
-
 .dfo-accordion {
   list-style: none;
   margin: 0;
