+++ src/Accordion/index.js
@@ -1,0 +1,1 @@
+export { default as Accordion } from "./Accordion";
