+++ src/Alertbox/Alertbox.scss
@@ -1,5 +1,3 @@
-@import "../_core/_index";
-
 .dfo-alertbox {
   background-color: $error;
   color: $textonprimary;
