+++ src/Alertbox/index.js
@@ -1,0 +1,1 @@
+export { default as Alertbox } from "./Alertbox";
