+++ src/Factbox/Factbox.scss
