--- src/components/reset.js
@@ -1,6 +1,0 @@
-import React from "react";
-import "./_core/_index.scss";
-
-export default function DFOReset() {
-  return <span>DFOReset: No need to add me to the page.</span>;
-}
