--- src/example-create-react-app-test/App.js
@@ -1,17 +1,0 @@
-import React, { Component } from "react";
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <p>This is not the Storybook you're looking for.</p>
-        <p>
-          See README.md or package.json for details about firing up the
-          Storybook.
-        </p>
-      </div>
-    );
-  }
-}
-
-export default App;
